Add authorization in every controller

+++ app/controllers/battles_controller.rb
@@ -1,8 +1,10 @@
 class BattlesController < ApplicationController
   def create
     @battle = Battle.new(battle_params)
+    authorize @battle
 
     if @battle.save
+      BattleMember.create(battle_id: @battle.id, member_id: battle_params[:owner_id], member_type: "User")
       respond_to do |format|
         format.html { redirect_to @battle }
         format.json { render json: @battle }
@@ -12,7 +14,9 @@ class BattlesController < ApplicationController
 
   def index
     @battles = Battle.all
+    authorize @battles
     @battle = Battle.new
+    authorize @battle, :new?
 
     respond_to do |format|
       format.html
@@ -22,16 +26,20 @@ class BattlesController < ApplicationController
 
   def show
     @battle = Battle.find(params[:id])
+    authorize @battle
+
     @challenge = @battle.challenges.new
     @battle_member = @battle.battle_members.new
   end
 
   def edit
     @battle = Battle.find(params[:id])
+    authorize @battle
   end
 
   def update
     @battle = Battle.find(params[:id])
+    authorize @battle
 
     if @battle.update_attributes(battle_params)
       respond_to do |format|
@@ -43,6 +51,7 @@ class BattlesController < ApplicationController
 
   def destroy
     @battle = Battle.find(params[:id])
+    authorize @battle
     
     if @battle.destroy
       respond_to do |format|
